third_party/WebKit/LayoutTests/inspector/geolocation-emulation-tests.html

<html>
<head>
<script src="../http/tests/inspector/inspector-test.js"></script>
<script>

function serializeGeolocationError(error) {
    var result = "Unknown error"
    switch (error.code)
    {
        case error.PERMISSION_DENIED:
            result = "Permission denied";
            break;
        case error.POSITION_UNAVAILABLE:
            result = "Position unavailable";
            break;
        case error.TIMEOUT:
            result = "Request timed out";
            break;
    }
    if (error.message)
        result += " (" + error.message + ")";
    return result;
}

function overrideGeolocation()
{
    function testSuccess(position)
    {
        if (position && position.coords)
            console.log("Latitude: " + position.coords.latitude + " Longitude: " + position.coords.longitude);
        else
            console.log("Unexpected error occured. Test failed.");
    }

    function testFailed(error)
    {
        console.log(serializeGeolocationError(error));
    }

    navigator.geolocation.getCurrentPosition(testSuccess, testFailed);
}

function overridenTimestampGeolocation()
{
    function testSuccess(position)
    {
        if ((new Date(position.timestamp)).toDateString() == (new Date()).toDateString())
            console.log("PASSED");
        else
            console.log("Unexpected error occured. Test failed.");
    }

    function testFailed(error)
    {
        console.log(serializeGeolocationError(error));
    }

    navigator.geolocation.getCurrentPosition(testSuccess, testFailed);
}

function test()
{
    InspectorTest.runTestSuite([
        function testGeolocationUnavailable(next)
        {
            PageAgent.setGeolocationOverride();
            InspectorTest.addConsoleSniffer(next);
            InspectorTest.evaluateInPage("overrideGeolocation()");
        },

        function testOverridenGeolocation(next)
        {
            PageAgent.setGeolocationOverride(50, 100, 95);
            InspectorTest.addConsoleSniffer(next);
            InspectorTest.evaluateInPage("overrideGeolocation()");
        },

        function testInvalidParam(next)
        {
            PageAgent.setGeolocationOverride(true, 100, 95);
            next();
        },

        function testInvalidGeolocation(next)
        {
            PageAgent.setGeolocationOverride(200, 300, 95);
            InspectorTest.addConsoleSniffer(next);
            InspectorTest.evaluateInPage("overrideGeolocation()");
        },

        function testTimestampOfOverridenPosition(next)
        {
            PageAgent.setGeolocationOverride(50, 100, 95);
            InspectorTest.addConsoleSniffer(next);
            InspectorTest.evaluateInPage("overridenTimestampGeolocation()");
        }
    ]);
}
</script>
</head>
<body onload="runTest()">
<p>
Tests that geolocation emulation with latitude and longitude works as expected.
</p>
</body>
</html>
