third_party/WebKit/LayoutTests/inspector/geolocation-emulation-tests-expected.txt

CONSOLE MESSAGE: line 37: Position unavailable
CONSOLE MESSAGE: line 30: Latitude: 50 Longitude: 100
CONSOLE MESSAGE: line 30: Latitude: 50 Longitude: 100
CONSOLE MESSAGE: line 48: PASSED
Tests that geolocation emulation with latitude and longitude works as expected.


Running: testGeolocationUnavailable

Running: testOverridenGeolocation

Running: testInvalidParam
error: Protocol Error: Invalid type of argument 'latitude' for method 'Page.setGeolocationOverride' call. It must be 'number' but it is 'boolean'.

Running: testInvalidGeolocation
error: Request with id = 30 failed. {"code":-32603,"message":"Invalid geolocation"}

Running: testTimestampOfOverridenPosition
